Add selenium test for q=null error
Also move test_ApiErrorNotLeaking to test_selenium.py
